+++ src/main/java/ma/salman/sbschool/repository/EtudiantRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface EtudiantRepository extends JpaRepository<Etudiant, Integer> {
+public interface EtudiantRepository extends JpaRepository<Etudiant, Long> {
     List<Etudiant> findByClasseId(Long classeId);
     List<Etudiant> findByParentId(Long parentId);
 }
